refactor: share Chronos-family export/test-data code via utils
Chronos 2 and Chronos-Bolt duplicated the masked instance-norm, the ONNX
export boilerplate, and nearly all of the synthetic test-data generation.
Move the shared parts into a new prep_models_utils.chronos subpackage (mirroring
prep_models_utils.astromer): masked_instance_norm, run_export, run_test_data.

Each model package now only keeps its own encoder wrapper (the part that
genuinely differs) plus a thin run_export/run_test_data shim. Behaviour is
unchanged — exports validate and the committed test-data parquets regenerate
byte-for-byte identically for all five models.
